Move test classes to correct package signatures
