+++ datacapture/src/test/java/com/google/android/fhir/datacapture/contrib/views/barcode/mlkit/md/camera/WorkflowModelTest.kt
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.utilities
+package com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.camera
 
 import androidx.test.core.app.ApplicationProvider
-import com.google.android.fhir.datacapture.contrib.views.barcode.mlkit.md.camera.WorkflowModel
 import com.google.common.truth.Truth.assertThat
 import java.util.HashSet
 import org.junit.Before
